pyproject.toml
[project]
name = "langchaindemo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.11",
    "langchain-openai>=1.1.11",
    "python-dotenv>=1.2.2",
    "tavily>=1.1.0",
]
